added all chat jsx into it's own component - ChatSearch
