+++ src/components/Chat/Chat.jsx
@@ -1,10 +1,9 @@
 import React, { useEffect, useState } from "react";
 import "./Chat.css";
 import { BiMessageRoundedAdd, BiSearch } from "react-icons/bi";
-import { CiMinimize1 } from "react-icons/ci";
-import { AiOutlineClose } from "react-icons/ai";
+
 import { getUsers } from "../../services/user";
-import InputUsers from "./InputUsers";
+import ChatSearch from "./ChatSearch";
 
 function Chat({ setToggleChat, setShowChat }) {
   const [users, setUsers] = useState(null);
@@ -20,13 +19,6 @@ function Chat({ setToggleChat, setShowChat }) {
     }
     getAllUsers();
   }, []);
-
-  function handleMinimize() {
-    setToggleChat(true);
-  }
-  function handleClose() {
-    setShowChat(false);
-  }
 
   function checkForUser(currentInput) {
     let result = [];
@@ -59,50 +51,15 @@ function Chat({ setToggleChat, setShowChat }) {
         </div>
       </div>
 
-      <div className="chat-main">
-        <div className="chat-main-nav">
-          <h3 className="chat-main-title">New Chat</h3>
-          <div className="chat-main-icons">
-            <CiMinimize1
-              className="chat-main-minimize"
-              onClick={handleMinimize}
-            />
-            <AiOutlineClose onClick={handleClose} />
-          </div>
-        </div>
-
-        <div className="chat-search-container">
-          <div>
-            <p className="chat-search-title">SEARCH</p>
-            <div className="chat-search-input-wrapper">
-              <div className="chat-search-input-field">
-                {recipients.map((recipient, i) => {
-                  return <span key={i}>{recipient}</span>;
-                })}
-                {/* <div> */}
-                {/* <BiSearch className="chat-search-icon"></BiSearch> */}
-                <input
-                  value={usernameInput}
-                  name="usernameInput"
-                  className="chat-search-input"
-                  type="text"
-                  placeholder="Type usernames"
-                  onChange={(e) => handleUsernameInput(e)}
-                />
-                {/* </div> */}
-              </div>
-              <InputUsers
-                users={showUsers}
-                handleSelectedUser={handleSelectedUser}
-              />
-            </div>
-          </div>
-          <div className="chat-search-buttons">
-            <button>Start Group</button>
-            <button>Cancel</button>
-          </div>
-        </div>
-      </div>
+      <ChatSearch
+        recipients={recipients}
+        usernameInput={usernameInput}
+        handleUsernameInput={handleUsernameInput}
+        showUsers={showUsers}
+        handleSelectedUser={handleSelectedUser}
+        setToggleChat={setToggleChat}
+        setShowChat={setShowChat}
+      />
     </div>
   );
 }
